% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combination.R
\name{Combination}
\alias{Combination}
\title{Combine the marginal p-values}
\usage{
Combination(
  y,
  X,
  taus = c(0.1, 0.25, 0.5, 0.75, 0.9),
  m = 3,
  test_num,
  method = "Chi"
)
}
\arguments{
\item{y}{n*1 vector, the observed outcome for n samples}

\item{X}{n*p matrix, the observed p covariates for n samples}

\item{taus}{k*1 vector, a grid of quantile levels; e.g., 0.5 for the median, 0.75 for the 3rd quartile; default is c(0.1, 0.25, 0.5, 0.75, 0.9)}

\item{m}{numeric variable, the order of B-spline function; default is 3}

\item{test_num}{a vector, representing the test corresponds to which covariate(s) in X.}

\item{method}{different method for calculating p-value: 'Chi' for large sample cases; 'Pearson' for small sample cases}
}
\value{
quantiles of a m*k matrix, each row is the estimated quantiles for each new case
}
\description{
Combine the marginal p-values
}
\details{
\itemize{
  \item Please choose 'Chi' or 'Pearson' for \code{method}, no other options.
  \item \code{taus} must be a subset or equal to the grid used to produce \code{input}.
}
}
